models/aprobaciones_prophet_2026/requirements.txt
pandas
numpy
prophet
scikit-learn
pyyaml
matplotlib
pytest